radius/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
strawberry-graphql==0.211.1
sqlalchemy==1.4.42
databases[mysql]==0.8.0
aiomysql==0.2.0
mysqlclient==2.2.0
python-dotenv==1.0.0 